Add price class and data structures for order book

+++ order.cpp
@@ -2,12 +2,11 @@
 #include "order.hpp"
 
 Order::Order(
-  unsigned int order_id,
-  std::string instrument,
-  double price, 
-  unsigned int count, 
-  std::string side, 
-  std::tm timestamp
-): order_id(order_id), instrument(instrument), price(price), count(count) , side(side), timestamp(timestamp) {
-
-}
+    unsigned int order_id, std::string instrument, unsigned int price,
+    unsigned int count, std::string side, std::tm timestamp)
+    : order_id(order_id),
+      instrument(instrument),
+      price(price),
+      count(count),
+      side(side),
+      timestamp(timestamp) {}
